requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
google-generativeai==0.3.1
pydantic-settings==2.0.3
python-dotenv==1.0.0